Move some things around (util and comms subpackages).
