+++ libensemble/tests/unit_tests/tofix__test_aaa_comms.py
@@ -13,7 +13,7 @@ import threading
 import queue
 
 import numpy as np
-import libensemble.comms as comms
+import libensemble.comms.comms as comms
 
 
 def test_qcomm_proc_terminate():
